Move application-wide actions to EvApplication
"Open" and "About" actions are independent of existing windows and can
live directly on the application object. This patch also adds those two
actions to the app menu.

Small regression: the comment field is not filled with information of
the backend used for the document in the current window anymore. It's
unclear which information should be shown there when multiple windows
with different backends are open.